+++ app/src/main/java/com/machiav3lli/backup/dbs/entity/AppInfo.kt
@@ -32,7 +32,7 @@ open class AppInfo : com.machiav3lli.backup.dbs.entity.PackageInfo {
 
     constructor(
         packageName: String,
-        packageLabel: String?,
+        packageLabel: String,
         versionName: String?,
         versionCode: Int,
         profileId: Int,