+++ app/src/main/java/com/machiav3lli/backup/pages/PermissionsPage.kt
@@ -65,8 +65,9 @@ import com.machiav3lli.backup.ui.compose.icons.Phosphor
 import com.machiav3lli.backup.ui.compose.icons.phosphor.ArrowRight
 import com.machiav3lli.backup.ui.compose.item.ElevatedActionButton
 import com.machiav3lli.backup.ui.compose.item.PermissionItem
-import com.machiav3lli.backup.ui.navigation.NavItem
+import com.machiav3lli.backup.ui.compose.item.TopBar
 import com.machiav3lli.backup.ui.item.Permission
+import com.machiav3lli.backup.ui.navigation.NavItem
 import com.machiav3lli.backup.utils.checkBatteryOptimization
 import com.machiav3lli.backup.utils.checkCallLogsPermission
 import com.machiav3lli.backup.utils.checkContactsPermission
@@ -148,6 +149,9 @@ fun PermissionsPage() {
     })
 
     Scaffold(
+        topBar = {
+            TopBar(title = stringResource(id = R.string.app_name)) {}
+        },
         bottomBar = {
             AnimatedVisibility(visible = permissionsList.isEmpty()) {
                 Row(
@@ -169,9 +173,9 @@ fun PermissionsPage() {
     ) { paddingValues ->
         LazyColumn(
             modifier = Modifier
+                .padding(paddingValues)
                 .blockBorder()
-                .fillMaxSize()
-                .padding(paddingValues),
+                .fillMaxSize(),
             verticalArrangement = Arrangement.spacedBy(8.dp),
             contentPadding = PaddingValues(8.dp)
         ) {
@@ -240,7 +244,7 @@ fun AppCompatActivity.showBatteryOptimizationDialog(powerManager: PowerManager?)
         .setMessage(R.string.ignore_battery_optimization_message)
         .setPositiveButton(R.string.dialog_approve) { _: DialogInterface?, _: Int ->
             val intent = Intent(Settings.ACTION_REQUEST_IGNORE_BATTERY_OPTIMIZATIONS)
-            intent.data = Uri.parse("package:" + packageName)
+            intent.data = Uri.parse("package:$packageName")
             try {
                 startActivity(intent)
                 persist_ignoreBatteryOptimization.value =