+++ app/src/main/java/com/machiav3lli/backup/ui/navigation/Host.kt
@@ -3,7 +3,9 @@ package com.machiav3lli.backup.ui.navigation
 import androidx.compose.animation.AnimatedVisibilityScope
 import androidx.compose.animation.fadeIn
 import androidx.compose.animation.fadeOut
+import androidx.compose.animation.slideInHorizontally
 import androidx.compose.animation.slideInVertically
+import androidx.compose.animation.slideOutHorizontally
 import androidx.compose.animation.slideOutVertically
 import androidx.compose.foundation.ExperimentalFoundationApi
 import androidx.compose.foundation.pager.PagerState
@@ -17,6 +19,7 @@ import androidx.navigation.activity
 import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
 import com.machiav3lli.backup.activities.PrefsActivityX
+import com.machiav3lli.backup.pages.MainPage
 import com.machiav3lli.backup.pages.PermissionsPage
 import com.machiav3lli.backup.pages.WelcomePage
 import com.machiav3lli.backup.preferences.ExportsPage
@@ -26,13 +29,10 @@ import com.machiav3lli.backup.preferences.persist_beenWelcomed
 import com.machiav3lli.backup.viewmodels.ExportsViewModel
 import com.machiav3lli.backup.viewmodels.LogViewModel
 
-@OptIn(ExperimentalFoundationApi::class)
 @Composable
 fun MainNavHost(
     modifier: Modifier = Modifier,
     navController: NavHostController,
-    pagerState: PagerState,
-    pages: List<NavItem>,
 ) {
     NavHost(
         modifier = modifier,
@@ -40,16 +40,14 @@ fun MainNavHost(
         startDestination = if (persist_beenWelcomed.value) NavItem.Permissions.destination
         else NavItem.Welcome.destination
     ) {
-        slideDownComposable(NavItem.Welcome.destination) {
+        slideInComposable(NavItem.Welcome.destination) {
             WelcomePage()
         }
-        composable(route = NavItem.Permissions.destination) {
+        slideInComposable(route = NavItem.Permissions.destination) {
             PermissionsPage()
         }
-        slideUpComposable(NavItem.Main.destination) {
-            SlidePager(
-                pagerState = pagerState,
-                pageItems = pages,
+        slideInComposable(NavItem.Main.destination) {
+            MainPage(
                 navController = navController
             )
         }
@@ -73,35 +71,35 @@ fun PrefsNavHost(
         navController = navController,
         startDestination = NavItem.Settings.destination,
     ) {
-        slideDownComposable(NavItem.Settings.destination) {
+        slideInComposable(NavItem.Settings.destination) {
             SlidePager(
                 pageItems = pages,
                 pagerState = pagerState,
                 navController = navController,
             )
         }
-        slideUpComposable(NavItem.Exports.destination) {
+        slideInComposable(NavItem.Exports.destination) {
             val viewModel = viewModels.find { it is ExportsViewModel } as ExportsViewModel
             ExportsPage(viewModel)
         }
-        slideUpComposable(NavItem.Logs.destination) {
+        slideInComposable(NavItem.Logs.destination) {
             val viewModel = viewModels.find { it is LogViewModel } as LogViewModel
             LogsPage(viewModel)
         }
-        slideUpComposable(NavItem.Terminal.destination) {
+        slideInComposable(NavItem.Terminal.destination) {
             TerminalPage()
         }
     }
 }
 
-fun NavGraphBuilder.slideUpComposable(
+fun NavGraphBuilder.slideInComposable(
     route: String,
     composable: @Composable (AnimatedVisibilityScope.(NavBackStackEntry) -> Unit),
 ) {
     composable(
         route,
-        enterTransition = { slideInVertically { height -> height } + fadeIn() },
-        exitTransition = { slideOutVertically { height -> -height } + fadeOut() }
+        enterTransition = { slideInHorizontally { width -> width } },
+        exitTransition = { slideOutHorizontally { width -> -width } },
     ) {
         composable(it)
     }