+++ CHANGELOG.md
@@ -1,6 +1,45 @@
 CHANGELOG
 =========
 
+8.3.4 (13.10.2023) +10 Commits & +10 Translations
+------------
+
+#### Function
+
+- Fix: Running schedules when app is not open
+- Fix: Missing stub for restorePackage in specials
+
+#### UI
+
+- Fix: NavBar overlap of sheets
+- Add: Animate navigation bar buttons
+
+#### UX
+
+- Fix: Crash on restarting app with open AppSheet
+
+
+8.3.3 (11.09.2023) +30 Commits & +30 Translations
+------------
+
+#### Function
+
+- Fix: Force stopping an app
+- Fix: Deleting schedule
+- Fix: Crash on context menu/Put
+- Fix: AutoLog after schedule in case of empty filtered list
+- Update: Abstract Main pages from Activity
+- Update: Generate Kotlin code using Room
+
+#### UI
+
+- Update: Pre-load all pages
+- Update: Improve animation of updated apps bar
+
+#### UX
+
+- Fix: Add workaround to NavBar overlap of sheets on specific Android versions
+
 8.3.2 (23.06.2023) +260 Commits & +70 Translations
 ------------